feat:내 정보 인터페이스 user.d.ts 로 이동, 유저 프로필 나타내는 Profile 인터페이스 추가
