+++ src/types/user.d.ts
@@ -1,3 +1,5 @@
+import { MyReview, Review } from './review';
+
 export interface User {
   _id: string;
   name: string;
@@ -8,4 +10,20 @@ export interface User {
   createdAt: string;
   favorites: { [key: string]: boolean };
   favoritesReview: { [key: string]: boolean };
+}
+
+export interface MyInformation {
+  comments: [];
+  favReviews: MyReview[];
+  favStories: [];
+  pageDescription: string;
+  pageKeywords: string;
+  pageTitle: string;
+  reviews: MyReview[];
+  stories: [];
+}
+
+export interface Profile {
+  users: string;
+  reviews: Review[];
 }
